fix: Fix memory leak in tests
When running unit tests, connections with database were not correctly closed, which resulted in memory leak and not killing all threads.
First, database connection was closed correctly, but Indexer shedules removal of unused jars from cache after 2 seconds,
which was creating new database and never closing it.

+++ metals/src/main/scala/scala/meta/internal/metals/Indexer.scala
@@ -362,7 +362,8 @@ final case class Indexer(
       sh.schedule(
         new Runnable {
           override def run(): Unit = {
-            tables.jarSymbols.deleteNotUsedTopLevels(usedJars.toArray)
+            if (tables.databaseExists())
+              tables.jarSymbols.deleteNotUsedTopLevels(usedJars.toArray)
           }
         },
         2,
